changing angular displacement variable to float, for greater accuracy

+++ CaroloCup.h
@@ -126,7 +126,7 @@ class Gyroscope {
 		int readRegister(int deviceAddress, byte address);
 		unsigned short _samplingRate;
 		static const unsigned short DEFAULT_GYRO_SAMPLING;
-		int _angularDisplacement;
+		float _angularDisplacement;
 		unsigned long _prevSample;
 };
